requirements.txt
fastapi
uvicorn
gunicorn
pandas
numpy
openai
langchain
azure-storage-blob==12.19.0
python-dotenv